Put validCountries in RobotServices
